requirements.txt
opencv-python
torch
numpy
sounddevice
SpeechRecognition
pyserial
requests